xbios: pass correct arg types per Alcyon prototypes
Alcyon OSBIND.O's xbios() is a variadic trap wrapper.  Callers must
push args at their DECLARED sizes to match what TOS reads from the
trap-frame stack:

    void Setscreen(VOIDP log, VOIDP phys, WORD mode)
    void Setpalette(VOIDP pal)
    void Dosound(VOIDP p)
    void Midiws(WORD n, VOIDP buf)
    LONG Giaccess(WORD data, WORD reg)

The port had been passing everything as (long) to force 4-byte
pushes, which mismatches WORD args by 2 bytes.  The old form
`Setscreen((long)g_dscp, -1L, -1L)` worked "by luck" for Setscreen
because -1L's upper half also reads as -1 WORD, but Alcyon's
compile-time-known arg-slot sizing then differs from what the
trap handler expects.

Fixes across 5 files, 7 Setscreen sites, 3 Setpalette, 1 Dosound,
1 Midiws, 6 Giaccess.  Also see the OSBIND.O text-segment
disassembly (16-byte identical shape for _xbios, _bios, _gemdos):

    23DF <saveret.L>    MOVE.L (A7)+, saveret   ; pop RA
    4E4E                TRAP #14                ; XBIOS
    2F39 <saveret.L>    MOVE.L saveret.L, -(A7) ; push RA back
    4E75                RTS                     ; caller cleans args

Note: this fix alone does not resolve the letter-writing action
freeze -- LCP.PRG still bus-errors at PC=$fd330c dereferencing a
bad pointer ($14e3bbe4) during the first prCh call.  The upstream
issue is elsewhere (likely a wrong prototype we haven't found yet
or an arg-slot corruption in one of the surrounding calls).

+++ csrc/sfx_irq.c
@@ -143,7 +143,7 @@ sf_irqp()
 
         /* Hand the sequence to the XBIOS Dosound player.  Runs from
            the VBL interrupt on the ST; no-op on the host. */
-        Dosound((long) g_sfDoB);
+        Dosound(g_sfDoB);
 
         /* Snapshot the 200 Hz counter for the countdown loop and
            compute the remaining ticks in game 8Hz units (Dosound
